+++ NEWS.md
@@ -82,6 +82,8 @@
  * New scenario option `targetRunnerTimeout`: Timeout in seconds of any `targetRunner` call (only applies to `target-runner` executables not to R functions).
  
  * `read_pcs_file()` now supports forbidden configurations.
+ 
+ * New function `get_instance_seed_pairs()` to get the pairs of instance and random seed used during the races.
 
 ## Fixes